Moved all wind calculation module and unit tests to a new subdirectory and fixed imports.  Unit tests run; CLI tests TBC. (#635)
