+++ bin/improver-nbhood
@@ -38,7 +38,7 @@ from improver.nbhood.nbhood import (
 from improver.nbhood.recursive_filter import RecursiveFilter
 from improver.utilities.load import load_cube
 from improver.utilities.save import save_netcdf
-from improver.wind_direction import WindDirection
+from improver.wind_calculations.wind_direction import WindDirection
 
 
 def main():
