+++ bin/improver-wind-direction
@@ -33,7 +33,7 @@
 
 from improver.argparser import ArgParser
 
-from improver.wind_direction import WindDirection
+from improver.wind_calculations.wind_direction import WindDirection
 from improver.utilities.load import load_cube
 from improver.utilities.save import save_netcdf
