+++ lib/improver/tests/nbhood/square_kernel/test_SquareNeighbourhood.py
@@ -41,7 +41,7 @@ from iris.tests import IrisTest
 import numpy as np
 
 from improver.nbhood.square_kernel import SquareNeighbourhood
-from improver.wind_direction import WindDirection
+from improver.wind_calculations.wind_direction import WindDirection
 from improver.tests.nbhood.nbhood.test_NeighbourhoodProcessing import (
     set_up_cube)
 from improver.utilities.warnings_handler import ManageWarnings
