+++ lib/improver/wind_calculations/wind_gust_diagnostic.py
